Move plotconfiguration to CandleChart
