+++ src/components/ftbot/CandleChart.vue
@@ -1,5 +1,19 @@
 <template>
   <div class="container-fluid d-flex flex-column align-items-stretch d-flex">
+    <b-modal
+      id="plotConfiguratorModal"
+      title="Plot Configurator"
+      ok-only
+      hide-backdrop
+      button-size="sm"
+    >
+      <PlotConfigurator :columns="datasetColumns" v-model="plotConfig" />
+    </b-modal>
+    <div class="col-mb-2 ml-auto position-relative">
+      <b-button class="mt-5" @click="showConfigurator" size="sm" title="Plot configurator">
+        &#9881;
+      </b-button>
+    </div>
     <div class="row flex-grow-1 chart-wrapper">
       <v-chart v-if="hasData" theme="dark" autoresize :options="chartOptions" />
     </div>
@@ -10,10 +24,12 @@
 import { Component, Vue, Prop, Watch } from 'vue-property-decorator';
 import ECharts from 'vue-echarts';
 import * as echarts from 'echarts';
-import { Trade, PairHistory, PlotConfig } from '@/store/types';
+import { Trade, PairHistory, PlotConfig, EMPTY_PLOTCONFIG } from '@/store/types';
 import randomColor from '@/shared/randomColor';
 import { roundTimeframe } from '@/shared/timemath';
 import { timestampms } from '@/shared/formatters';
+import PlotConfigurator from '@/components/ftbot/PlotConfigurator.vue';
+import { loadCustomPlotConfig } from '@/shared/storage';
 
 import 'echarts';
 
@@ -26,7 +42,7 @@ const downColor = '#ec0000';
 const downBorderColor = '#8A0000';
 
 @Component({
-  components: { 'v-chart': ECharts },
+  components: { 'v-chart': ECharts, PlotConfigurator },
 })
 export default class CandleChart extends Vue {
   @Prop({ required: true }) readonly pair!: string;
@@ -39,7 +55,7 @@ export default class CandleChart extends Vue {
 
   @Prop({ required: true }) readonly dataset!: PairHistory;
 
-  @Prop({ required: false }) readonly plotConfig!: PlotConfig;
+  plotConfig: PlotConfig = { ...EMPTY_PLOTCONFIG };
 
   // Only recalculate buy / sell data if necessary
   signalsCalculated = false;
@@ -47,6 +63,14 @@ export default class CandleChart extends Vue {
   buyData = [] as Array<number>[];
 
   sellData = [] as Array<number>[];
+
+  mounted() {
+    this.plotConfig = loadCustomPlotConfig();
+  }
+
+  showConfigurator() {
+    this.$bvModal.show('plotConfiguratorModal');
+  }
 
   @Watch('pair')
   pairChanged() {
@@ -61,6 +85,10 @@ export default class CandleChart extends Vue {
   @Watch('dataset')
   datasetChanged() {
     this.signalsCalculated = false;
+  }
+
+  get datasetColumns() {
+    return this.dataset ? this.dataset.columns : [];
   }
 
   get hasData() {
