+++ src/views/Graphs.vue
@@ -10,33 +10,17 @@
       <div class="col-mb-2">
         <b-form-select :options="whitelist" v-model="pair" @change="refresh"> </b-form-select>
       </div>
-      <div class="col-mb-2">
-        <b-checkbox v-model="strategyPlotConfig">Use strategy plot_config</b-checkbox>
-      </div>
-      <div class="col-mb-2 ml-5" v-if="!strategyPlotConfig">
-        <b-button @click="showConfigurator">Show configurator</b-button>
-      </div>
     </div>
     <div class="mt-2" v-if="historicView">
       <TimeRangeSelect v-model="timerange"></TimeRangeSelect>
     </div>
 
-    <b-modal
-      id="plotConfiguratorModal"
-      title="Plot Configurator"
-      ok-only
-      hide-backdrop
-      button-size="sm"
-    >
-      <PlotConfigurator :columns="datasetColumns" v-model="customPlotConfig" />
-    </b-modal>
     <div class="row">
       <CandleChart
         :pair="pair"
         :timeframe="timeframe"
         :timeframems="timeframems"
         :dataset="dataset"
-        :plotConfig="selectedPlotConfig"
         :trades="trades"
       >
       </CandleChart>
@@ -48,14 +32,12 @@
 import { Component, Vue } from 'vue-property-decorator';
 import { namespace } from 'vuex-class';
 import CandleChart from '@/components/ftbot/CandleChart.vue';
-import PlotConfigurator from '@/components/ftbot/PlotConfigurator.vue';
 import TimeRangeSelect from '@/components/ftbot/TimeRangeSelect.vue';
-import { PlotConfig, EMPTY_PLOTCONFIG, PairCandlePayload, PairHistoryPayload } from '@/store/types';
-import { loadCustomPlotConfig } from '@/shared/storage';
+import { PairCandlePayload, PairHistoryPayload } from '@/store/types';
 
 const ftbot = namespace('ftbot');
 @Component({
-  components: { CandleChart, PlotConfigurator, TimeRangeSelect },
+  components: { CandleChart, TimeRangeSelect },
 })
 export default class Graphs extends Vue {
   pair = 'XRP/USDT';
@@ -64,27 +46,17 @@ export default class Graphs extends Vue {
 
   timeframems = 300000;
 
-  strategyPlotConfig = false;
-
-  plotOption = 'main_plot';
-
   historicView = false;
 
   strategy = '';
 
   timerange = '';
 
-  // Custom plot config - manually changed by user.
-  // eslint-disable-next-line @typescript-eslint/camelcase
-  customPlotConfig: PlotConfig = { ...EMPTY_PLOTCONFIG };
-
   @ftbot.State candleData;
 
   @ftbot.State history;
 
   @ftbot.State whitelist;
-
-  @ftbot.State plotConfig;
 
   @ftbot.State trades;
 
@@ -97,18 +69,10 @@ export default class Graphs extends Vue {
   @ftbot.Action
   public getWhitelist;
 
-  @ftbot.Action
-  public getPlotConfig;
-
   mounted() {
     this.getWhitelist();
     this.refresh();
     // eslint-disable-next-line @typescript-eslint/camelcase
-    this.customPlotConfig = loadCustomPlotConfig();
-  }
-
-  get selectedPlotConfig() {
-    return this.strategyPlotConfig ? this.plotConfig : this.customPlotConfig;
   }
 
   get dataset() {
@@ -116,14 +80,6 @@ export default class Graphs extends Vue {
       return this.history[`${this.pair}__${this.timeframe}`];
     }
     return this.candleData[`${this.pair}__${this.timeframe}`];
-  }
-
-  get datasetColumns() {
-    return this.dataset ? this.dataset.columns : [];
-  }
-
-  showConfigurator() {
-    this.$bvModal.show('plotConfiguratorModal');
   }
 
   refresh() {
@@ -137,7 +93,6 @@ export default class Graphs extends Vue {
     } else {
       this.getPairCandles({ pair: this.pair, timeframe: this.timeframe, limit: 500 });
     }
-    this.getPlotConfig();
   }
 }
 </script>
